Search and sort + some Error messages is added for Courses and Subjects

+++ LMS-Project/Repositories/SubjectsRepository.cs
@@ -8,7 +8,7 @@ using System.Web;
 
 namespace LMS_Project.Repositories
 {
-    public class SubjectsRepository:IDisposable
+    public class SubjectsRepository : IDisposable
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
@@ -30,7 +30,7 @@ namespace LMS_Project.Repositories
         public bool Add(Subject subject)
         {
             var _subjects = this.Subjects().Where(s => s.Name.ToLower() == subject.Name.ToLower());
-            if(_subjects.Count()!=0)
+            if (_subjects.Count() != 0)
             {
                 _subjects = null;
                 return false;
@@ -42,8 +42,8 @@ namespace LMS_Project.Repositories
 
         public bool Edit(Subject subject)
         {
-            var _subjects = this.Subjects().Where(s => s.Name.ToLower() == subject.Name.ToLower() &&subject.ID!=s.ID);
-            if(_subjects.Count()!=0)
+            var _subjects = this.Subjects().Where(s => s.Name.ToLower() == subject.Name.ToLower() && subject.ID != s.ID);
+            if (_subjects.Count() != 0)
             {
                 _subjects = null;
                 return false;
@@ -53,15 +53,20 @@ namespace LMS_Project.Repositories
             return true;
         }
 
-        public void Delete(int? id)
+        public bool Delete(int? id)
         {
             Subject subject = Subject(id);
 
             if (subject != null)
             {
-                db.Subjects.Remove(subject);
-                SaveChanges();
+                if (subject.Courses.Where(c => c.Documents.Count() == 0 && c.SubjectID == id).Count() == 0)
+                {
+                    db.Subjects.Remove(subject);
+                    SaveChanges();
+                    return true;
+                }
             }
+            return false;
         }
         public List<Subject> AvaibleSubjects(string user)
         {
@@ -81,7 +86,7 @@ namespace LMS_Project.Repositories
         {
             db.SaveChanges();
         }
-        
+
         #region IDisposable Support
         private bool disposedValue = false;
 
@@ -99,5 +104,5 @@ namespace LMS_Project.Repositories
             Dispose(true);
         }
         #endregion
-    }  
+    }
 }
